services/telegram-bot/requirements.txt
aiogram==3.13.1
aiogram-dialog==2.3.1
httpx==0.27.2
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
redis==5.2.0
pydantic==2.10.2 